finish edit ActivityListCard
