--- src/Components/ActivityListCard/ActivityListCard.css
@@ -1,51 +1,0 @@
-
-.container{
-    background-color: salmon;
-    height: 861.23px;
-    width: 715px;
-    position: relative;
-
-}
-
-
-.posts{
-    width: 700px;
-    height: 640px;
-    border-radius: 20px;
-    overflow:scroll;
-}
-
-.posts ul{
-    width: 643px;
-    height: 145.12px;
-    list-style: none;
-    background-color: aquamarine;
-    margin: 20px 0 ;
-    border-radius: 20px;
-    padding: 15px;
-    box-shadow: rgba(100, 100, 111, 0.2) 5px 7px 10px;
-}
-
-
-.pagegination{
-    width: 100%;
-    list-style: none;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    position: absolute;
-    top: 85%;
-}
-
-
-
-
-.pagegination li{
-    padding: 5px;
-    font-size: 1.2em;
-}
-
-.pagegination li:hover{
-    cursor: pointer;
-    background-color: antiquewhite;
-}
