+++ src/Components/Page1/Activity-list/ActivityList.css
@@ -1,5 +1,7 @@
 * {
-    margin: 0px;
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
 }
 
 .act-container {
@@ -10,6 +12,7 @@
     background: #fd9d9dcc;
     border-radius: 25px;
     margin: 70px 100px 0px 20px;
+    position: relative;
 }
 
 .act-container h3 {
@@ -23,27 +26,17 @@
 
 .head-content{
     padding-top: 20px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
     margin: 20px 40px 0;
 }
+
 
 .head-content h1 {
     font-size: 24px;
     font-family: 'Varela Round', sans-serif;
     font-style: normal;
     color:  #572e2ec0;
-}
-
-.head-content img{
-    width: 43px;
-    height: 46px;
-    background-color: #F1F1F1;
-    margin: 0 10px;
-    padding: 10px;
-    border-radius: 25px;
-    box-shadow: 5px 5px 4px rgb(0 0 0 / 0.2);
+    box-shadow: 0 25px 35px -20px rgb(0 0 0 / 15%);
+    padding: 15px;
 
 }
