+++ src/Components/Page1/Activity-list/ActivityList.jsx
@@ -1,21 +1,35 @@
 import "./ActivityList.css";
 import ActivityListCard from "./ActivityListCard";
 import AddActivity from "./AddActivity";
+import { useState, useEffect } from "react";
+import axios from "axios";
 
-import clickNext from './images/next.png';
-import clickPrev from './images/prev.png';
 
 const ActivityList = () => {
+  const [activity, setActivity] = useState([]);
+
+  //Fetch data from database to show at card
+  const fetchData = () => {
+    axios
+    .get(`https://jsonplaceholder.typicode.com/posts`)
+    .then((res) => {
+      setActivity(res.data)
+    })
+    .catch((err) => {
+      alert(err)
+    })
+  }
+
+  useEffect(() => {
+    fetchData()
+  }, [])
+  
   return (
     <div className="act-container">
         <div className="head-content">
-                <h1> Activity List</h1>
-                <div className="head-btn">
-                    <a className="btn-left" href="#"><img src= { clickPrev } alt="previous" /></a>
-                    <a className="btn-right" href="#"><img src= { clickNext } alt="next" /></a>
-                </div>
-            </div>
-        <ActivityListCard />
+          <h1> Activity List</h1>
+        </div>
+        <ActivityListCard data= {activity}/>
         <AddActivity />
     </div>
   );
