+++ src/Components/Page1/Activity-list/AddActivity.css
@@ -6,7 +6,8 @@
     display: flex;
     justify-content: center;
     height: 60px;
-    margin-bottom: 30px;
+    /* margin-bottom: 30px; */
+    margin: 90px 0 20px 0;
 }
 
 .AddAct h1 {
