+++ src/Components/Page1/Activity-list/AddActivity.jsx
@@ -1,15 +1,12 @@
-import { Link } from "react-router-dom";
 import "./AddActivity.css";
 
 
 const AddActivity = () => {
   return (
     <div className="AddAct">
-      <Link to="/create">
       <button>
         <h1>Add Activity</h1>
       </button>
-      </Link>
     </div>
   );
 };
